Extract exercises into modules
